Rewrite Switch using TS
