+++ src/elements/Switch/index.ts
@@ -1,0 +1,1 @@
+export { default as Switch } from './Switch';
